Feature/UI manager init

+++ Engine/Mesh/Private/ResourceManager.cpp
@@ -1,7 +1,7 @@
 #include "pch.h"
 #include "Mesh/Public/ResourceManager.h"
 #include "Mesh/Public/VertexDatas.h"
-#include "Render/Public/Renderer.h"
+#include "Render/Renderer/Public/Renderer.h"
 
 IMPLEMENT_SINGLETON(UResourceManager)
 
@@ -26,29 +26,39 @@ void UResourceManager::Initialize()
 	VertexDatas.emplace(EPrimitiveType::GizmoB, &VerticesGizmoBlue);
 
 	//TArray.GetData(), TArray.Num()*sizeof(FVertexSimple), TArray.GetTypeSize()
-	Vertexbuffers.emplace(EPrimitiveType::Cube, Renderer.CreateVertexBuffer(VerticesCube.data(), VerticesCube.size()*sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::Sphere, Renderer.CreateVertexBuffer(VerticesSphere.data(), VerticesSphere.size() * sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::Triangle, Renderer.CreateVertexBuffer(VerticesTriangle.data(), VerticesTriangle.size() * sizeof(FVertex)));
+	Vertexbuffers.emplace(EPrimitiveType::Cube, Renderer.CreateVertexBuffer(
+		                      VerticesCube.data(), static_cast<int>(VerticesCube.size()) * sizeof(FVertex)));
+	Vertexbuffers.emplace(EPrimitiveType::Sphere, Renderer.CreateVertexBuffer(
+		                      VerticesSphere.data(), static_cast<int>(VerticesSphere.size() * sizeof(FVertex))));
+	Vertexbuffers.emplace(EPrimitiveType::Triangle, Renderer.CreateVertexBuffer(
+		                      VerticesTriangle.data(), static_cast<int>(VerticesTriangle.size() * sizeof(FVertex))));
 
-	Vertexbuffers.emplace(EPrimitiveType::LineR, Renderer.CreateVertexBuffer(VerticesLineRed.data(), VerticesLineRed.size() * sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::LineG, Renderer.CreateVertexBuffer(VerticesLineGreen.data(), VerticesLineGreen.size() * sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::LineB, Renderer.CreateVertexBuffer(VerticesLineBlue.data(), VerticesLineBlue.size() * sizeof(FVertex)));
+	Vertexbuffers.emplace(EPrimitiveType::LineR, Renderer.CreateVertexBuffer(
+		                      VerticesLineRed.data(), static_cast<int>(VerticesLineRed.size() * sizeof(FVertex))));
+	Vertexbuffers.emplace(EPrimitiveType::LineG, Renderer.CreateVertexBuffer(
+		                      VerticesLineGreen.data(), static_cast<int>(VerticesLineGreen.size() * sizeof(FVertex))));
+	Vertexbuffers.emplace(EPrimitiveType::LineB, Renderer.CreateVertexBuffer(
+		                      VerticesLineBlue.data(), static_cast<int>(VerticesLineBlue.size() * sizeof(FVertex))));
 
-	Vertexbuffers.emplace(EPrimitiveType::GizmoR, Renderer.CreateVertexBuffer(VerticesGizmoRed.data(), VerticesGizmoRed.size() * sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::GizmoG, Renderer.CreateVertexBuffer(VerticesGizmoGreen.data(), VerticesGizmoGreen.size() * sizeof(FVertex)));
-	Vertexbuffers.emplace(EPrimitiveType::GizmoB, Renderer.CreateVertexBuffer(VerticesGizmoBlue.data(), VerticesGizmoBlue.size() * sizeof(FVertex)));
+	Vertexbuffers.emplace(EPrimitiveType::GizmoR, Renderer.CreateVertexBuffer(
+		                      VerticesGizmoRed.data(), static_cast<int>(VerticesGizmoRed.size() * sizeof(FVertex))));
+	Vertexbuffers.emplace(EPrimitiveType::GizmoG, Renderer.CreateVertexBuffer(
+		                      VerticesGizmoGreen.data(),
+		                      static_cast<int>(VerticesGizmoGreen.size() * sizeof(FVertex))));
+	Vertexbuffers.emplace(EPrimitiveType::GizmoB, Renderer.CreateVertexBuffer(
+		                      VerticesGizmoBlue.data(), static_cast<int>(VerticesGizmoBlue.size() * sizeof(FVertex))));
 
-	NumVertices.emplace(EPrimitiveType::Cube, VerticesCube.size());
-	NumVertices.emplace(EPrimitiveType::Sphere, VerticesSphere.size());
-	NumVertices.emplace(EPrimitiveType::Triangle, VerticesTriangle.size());
+	NumVertices.emplace(EPrimitiveType::Cube, static_cast<UINT>(VerticesCube.size()));
+	NumVertices.emplace(EPrimitiveType::Sphere, static_cast<UINT>(VerticesSphere.size()));
+	NumVertices.emplace(EPrimitiveType::Triangle, static_cast<UINT>(VerticesTriangle.size()));
 
-	NumVertices.emplace(EPrimitiveType::LineR, VerticesLineRed.size());
-	NumVertices.emplace(EPrimitiveType::LineG, VerticesLineGreen.size());
-	NumVertices.emplace(EPrimitiveType::LineB, VerticesLineBlue.size());
+	NumVertices.emplace(EPrimitiveType::LineR, static_cast<UINT>(VerticesLineRed.size()));
+	NumVertices.emplace(EPrimitiveType::LineG, static_cast<UINT>(VerticesLineGreen.size()));
+	NumVertices.emplace(EPrimitiveType::LineB, static_cast<UINT>(VerticesLineBlue.size()));
 
-	NumVertices.emplace(EPrimitiveType::GizmoR, VerticesGizmoRed.size());
-	NumVertices.emplace(EPrimitiveType::GizmoG, VerticesGizmoGreen.size());
-	NumVertices.emplace(EPrimitiveType::GizmoB, VerticesGizmoBlue.size());
+	NumVertices.emplace(EPrimitiveType::GizmoR, static_cast<UINT>(VerticesGizmoRed.size()));
+	NumVertices.emplace(EPrimitiveType::GizmoG, static_cast<UINT>(VerticesGizmoGreen.size()));
+	NumVertices.emplace(EPrimitiveType::GizmoB, static_cast<UINT>(VerticesGizmoBlue.size()));
 }
 
 void UResourceManager::Release()
